phylib_uart: Add inter-byte delay on tx
Limit tx data traffic comply with low traffic capabilties seen for
some SoCs with LPUART, such as i.MX91, i.MX93 and i.MX95.

This fixes the program flow being broken due to FIFO overruns on
the target.

The delay of 20ms was tweaked based on tests with i.MX91 and i.MX93.

Signed-off-by: Alexander Herfurtner <alexander.h@variscite.com>
Tested-by: Natalia Kovalenko <natasha.k@variscite.com>